Restructure with a separate lib.
